refactor: ProjectDetail, ProjectTabs 날짜 함수 util.ts 에서 참조하는 형태로 변경
